+++ .eslintrc.json
@@ -1,3 +1,3 @@
 {
-	"extends": "next/core-web-vitals"
+	"extends": ["next/core-web-vitals", "plugin:storybook/recommended"]
 }
