+++ components/SiteFooter/SiteFooter.stories.tsx
@@ -1,0 +1,10 @@
+import { SiteFooter } from './SiteFooter';
+
+export default {
+	title: 'Components/SiteFooter',
+	component: SiteFooter,
+};
+
+export const Default = () => {
+	return <SiteFooter />;
+};
