+++ components/SiteHeader/SiteHeader.stories.tsx
@@ -2,6 +2,7 @@ import { SiteHeader } from './SiteHeader';
 
 export default {
 	title: 'Components/SiteHeader',
+	component: SiteHeader,
 };
 
 export const Default = () => {
