+++ package.json
@@ -70,7 +70,8 @@
 		"facepaint": "^1.2.1",
 		"next": "^12.1.6",
 		"react": "18.2.0",
-		"react-dom": "18.2.0"
+		"react-dom": "18.2.0",
+		"storybook-addon-next-router": "^4.0.1"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.20.5",
