+++ pages/index.tsx
@@ -1,5 +1,4 @@
 import type { NextPage } from 'next';
-import { AppLayout } from '../components/AppLayout';
 import { DocumentTitle } from '../components/DocumentTitle';
 import { HomePage } from '../components/HomePage';
 
@@ -7,9 +6,7 @@ const Home: NextPage = () => {
 	return (
 		<>
 			<DocumentTitle title="Home" />
-			<AppLayout>
-				<HomePage />
-			</AppLayout>
+			<HomePage />
 		</>
 	);
 };
